+++ db/migrate/20191108101914_add_num_entry_to_analyses.rb
@@ -1,0 +1,5 @@
+class AddNumEntryToAnalyses < ActiveRecord::Migration[5.0]
+  def change
+    add_column :analyses, :num_entry, :integer, after: :num_feature
+  end
+end